ingestion/requirements.txt
langchain
langchain-pinecone
pinecone
python-dotenv
sentence-transformers
langchain-huggingface
